Refactor: extract new SlaRequestUtils
